+++ pkg/sidecar/paths_posix_test.go
@@ -6,50 +6,6 @@ import (
 	"testing"
 )
 
-// TestPathIsVolumeMountPoint tests PathIsVolumeMountPoint.
-func TestPathIsVolumeMountPoint(t *testing.T) {
-	// Define test cases.
-	tests := []struct {
-		path               string
-		expectedVolumeness bool
-		expectedVolumeName string
-	}{
-		{"", false, ""},
-		{"/", false, ""},
-		{"/volume", false, ""},
-		{"/volume/", false, ""},
-		{"/volume/fake", false, ""},
-		{"/whatever", false, ""},
-		{"/whatever ", false, ""},
-		{"/whatever /", false, ""},
-		{"/volumes/", false, ""},
-		{"/volumes/ ", true, " "},
-		{"/volumes/name", true, "name"},
-		{"/volumes/name/", false, ""},
-		{"/volumes/my volume", true, "my volume"},
-		{"/volumes/ volume", true, " volume"},
-		{"/volumes/name/sub", false, ""},
-		{"/volumes/my volume/sub", false, ""},
-		{"/volumes/name/sub/second", false, ""},
-		{"/volumes/my volume/sub/second sub", false, ""},
-	}
-
-	// Process test cases.
-	for i, test := range tests {
-		volumeness, volumeName := PathIsVolumeMountPoint(test.path)
-		if volumeness != test.expectedVolumeness {
-			t.Errorf("test case %d: volumeness does not match expected: %t != %t",
-				i, volumeness, test.expectedVolumeness,
-			)
-		}
-		if volumeName != test.expectedVolumeName {
-			t.Errorf("test case %d: volume name does not match expected: %s != %s",
-				i, volumeName, test.expectedVolumeName,
-			)
-		}
-	}
-}
-
 // TestVolumeMountPointForPath tests VolumeMountPointForPath.
 func TestVolumeMountPointForPath(t *testing.T) {
 	// Define test cases.
