+++ pkg/sidecar/paths_windows_test.go
@@ -1,0 +1,3 @@
+package sidecar
+
+// TODO: Implement tests.
